@slaren 2023-05-06
This is good if it increases performance by 14%, but your results seem to contradict the tests from @dfyz [here](https://github.com/ggerganov/llama.cpp/pull/1221#issuecomment-1529166253), that showed that increasing occupancy does not always lead to increased performance.

Ideall though, we would use `cudaOccupancyMaxPotentialBlockSize ` to determine the block size, and add a check in the kernels to ignore the extra elements if it is not perfectly divisible.