@dfyz 2023-05-06
> my hardware is much slower so the dequantization is presumably slow enough that there is room for improvement

This is very encouraging! I think that simply changing the block size (as opposed to what I did in the #1221) is a very non-invasive change, so as long as it results in speeding up kernels at least on some GPUs, it should be merged.

I actually tried something similar to your PR [here](https://github.com/ggerganov/llama.cpp/pull/1221#issuecomment-1529166253) (see the "hacky patch"), but the `dequantize_block_q4_0` *kernel* actually got significantly slower on RTX 3070, so I abandoned this approach. This was with 768 threads (the result of `cudaOccupancyMaxPotentialBlockSize`) and with an additional check to make sure we ignore the extra elements if the number of blocks is not divisible by 768.

As I said, that was a hacky patch, so it's possible that I got something wrong. If someone with a relatively recent GPU can confirm that this PR doesn't slow down the kernels, I think we can proceed with increasing the block size for all kernels and merging the PR (thought it's ultimately up to @slaren).