@slaren 2023-05-06
> If someone with a relatively recent GPU can confirm that this PR doesn't slow down the kernels, I think we can proceed with increasing the block size for all kernels and merging the PR (thought it's ultimately up to @slaren).

I agree, and even though I appreciate the gesture, I am just a voice here, I don't "own" any code.