@SlyEcho 2023-05-07
FYI, getting 15% perf increase on my Vega64 as well.