@JohannesGaessler 2023-05-07
With `cudaOccupancyMaxPotentialBlockSize` the kernel size on my GTX 1070 is set to 1024. With 33b q4_0 the speedup for prompt processing is now 15%. Someone with a relatively recent GPU please confirm that this PR at least does not give them a performance regression.