@SlyEcho 2023-05-07
I did some tests, first line is the old (master) then second is the patch in this PR. Running on the perplexity tool with the first 103 lines from wiki.test.raw, 8 threads, no mmap, LLaMa 7B.

RX 6900XT, Linux 5.15, ROCm 5.2.3
```
llama_print_timings: prompt eval time = 25154.07 ms /  6144 tokens (    4.09 ms per token)
llama_print_timings: prompt eval time = 25642.30 ms /  6144 tokens (    4.17 ms per token)
```

RTX 2080 Ti, Linux 5.8, CUDA 11.4
```
llama_print_timings: prompt eval time = 27952.36 ms /  6144 tokens (    4.55 ms per token)
llama_print_timings: prompt eval time = 29715.57 ms /  6144 tokens (    4.84 ms per token)
```

RTX 3090, Linux 5.15, CUDA 11.4:
```
llama_print_timings: prompt eval time = 20707.82 ms /  6144 tokens (    3.37 ms per token)
CUDA error 700 at ~/src/llama.cpp/ggml-cuda.cu:682: an illegal memory access was encountered
```

The machines are different so not comparable directly to each other.