@JohannesGaessler 2023-05-07
I think I passed the wrong argument for determining the data limit. Can you re-run the test for the 3090 and check whether it works now?