@SlyEcho 2023-05-07
RX 6900XT:
```
llama_print_timings: prompt eval time = 25508.54 ms /  6144 tokens (    4.15 ms per token)
llama_print_timings: prompt eval time = 26084.42 ms /  6144 tokens (    4.25 ms per token)
```

RTX 2080 Ti:
```
llama_print_timings: prompt eval time = 28185.34 ms /  6144 tokens (    4.59 ms per token)
llama_print_timings: prompt eval time = 29794.24 ms /  6144 tokens (    4.85 ms per token)
```

RTX 3090:
```
llama_print_timings: prompt eval time = 20290.44 ms /  6144 tokens (    3.30 ms per token)
llama_print_timings: prompt eval time = 18930.16 ms /  6144 tokens (    3.08 ms per token)
```

Same setup as before. There may be something about the older software versions. I can't update anything since I don't own them, I only have access. Also, other users are running stuff on these machines, too but today seems quieter.

Sidenote, I should automate this somehow...