@JohannesGaessler 2023-05-07
It seems performance isn't consistently better/worse... should we make larger CUDA blocks a compilation option?