@SlyEcho 2023-05-07
I'd like someone else test as well, maybe on CUDA 12 it is better?