@slaren 2023-05-07
RTX 3080, CUDA 12.1

```
Master: llama_print_timings: prompt eval time = 25057.53 ms /  6144 tokens (    4.08 ms per token)
PR:     llama_print_timings: prompt eval time = 26587.40 ms /  6144 tokens (    4.33 ms per token)
```