@SlyEcho 2023-05-08
I tried some hacking:

It seems like the early return is liked better by the GPU:

```cuda
static __global__ void dequantize_block_q4_2(const void * vx, float * y, int k) {
    const block_q4_2 * x = (const block_q4_2 *) vx;

    const int i = blockIdx.x*blockDim.x + threadIdx.x;

    if (i >= k) {
        return;
    }

    const float d = x[i].d;
    const uint8_t * pp = x[i].qs;

    for (int l = 0; l < QK4_2; l += 2) {
        const uint8_t vi = pp[l/2];

        const int8_t vi0 = vi & 0xf;
        const int8_t vi1 = vi >> 4;

        const float v0 = (vi0 - 8)*d;
        const float v1 = (vi1 - 8)*d;

        y[i*QK4_2 + l + 0] = v0;
        y[i*QK4_2 + l + 1] = v1;
    }
}
```

Also, only check the max block size once, and only use it if it's lower than 256 like the original PR:
```cuda
static void dequantize_row_q4_2_cuda(const void * vx, float * y, int k, cudaStream_t stream) {
    const int nb = k / QK4_2;
    static int block_size = -1;
    if (block_size == -1) {
        int min_grid_size = -1;
        CUDA_CHECK(cudaOccupancyMaxPotentialBlockSize(&min_grid_size, &block_size, dequantize_block_q4_2, 0, 0));
        if (block_size > 256) block_size = 256;
    }
    int grid_size = (nb + block_size - 1) / block_size; // Round up.
    dequantize_block_q4_2<<<grid_size, block_size, 0, stream>>>(vx, y, nb);
}
```

With that I get to pretty much the same performance as with master, but I didn't test too exhaustively.