@SlyEcho 2023-05-08
This number, 256, is just magic, maybe it should be an option in the CMake file so the user can set it? Command line?

Although, I also tried 128 and 64 and they were slower.