@JohannesGaessler 2023-05-08
I already posted this in the PR about faster quantization kernels, but I also quickly tried implementing a faster quantization kernel: https://github.com/ggerganov/llama.cpp/pull/1221#issuecomment-1537540503 . With that kernel the number of blocks is 32 (16 for q4_2) times higher. I would intuitively assume that larger blocks would then give you better performance because there are fewer idle threads and (I think?) the GPU will have more opportunity for latency hiding.