@JohannesGaessler 2023-05-08
Maybe we should just wait for the quantization changes and then simultaneously optimize kernels and block size.